+++ examples/src/Console.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE CPP #-}
-module Teletype where
+module Console where
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative (pure)
@@ -16,27 +16,27 @@ import Control.Monad.Freer.Internal
 -------------------------------------------------------------------------------
                           -- Effect Model --
 -------------------------------------------------------------------------------
-data Teletype s where
-  PutStrLn    :: String -> Teletype ()
-  GetLine     :: Teletype String
-  ExitSuccess :: Teletype ()
+data Console s where
+  PutStrLn    :: String -> Console ()
+  GetLine     :: Console String
+  ExitSuccess :: Console ()
 
-putStrLn' :: Member Teletype r => String -> Eff r ()
+putStrLn' :: Member Console r => String -> Eff r ()
 putStrLn' = send . PutStrLn
 
-getLine'  :: Member Teletype r => Eff r String
+getLine'  :: Member Console r => Eff r String
 getLine' = send GetLine
 
-exitSuccess' :: Member Teletype r => Eff r ()
+exitSuccess' :: Member Console r => Eff r ()
 exitSuccess' = send ExitSuccess
 
 -------------------------------------------------------------------------------
                      -- Effectful Interpreter Simple --
 -------------------------------------------------------------------------------
-runTeletype :: Eff '[Teletype, IO] w -> IO w
-runTeletype req = runM (handleRelay pure go req)
+runConsole :: Eff '[Console, IO] w -> IO w
+runConsole req = runM (handleRelay pure go req)
   where
-   go :: Teletype v -> Arr '[IO] v w -> Eff '[IO] w
+   go :: Console v -> Arr '[IO] v w -> Eff '[IO] w
    go (PutStrLn msg) q = send (putStrLn msg) >>= q
    go GetLine q = send getLine >>= q
    go ExitSuccess q = send exitSuccess >>= q
@@ -44,12 +44,12 @@ runTeletype req = runM (handleRelay pure go req)
 -------------------------------------------------------------------------------
                         -- Pure Interpreter Simple --
 -------------------------------------------------------------------------------
-runTeletypePure :: [String] -> Eff '[Teletype] w -> [String]
-runTeletypePure inputs req =
+runConsolePure :: [String] -> Eff '[Console] w -> [String]
+runConsolePure inputs req =
   reverse . snd $ run (handleRelayS (inputs, []) (\s _ -> pure s) go req)
   where
     go :: ([String], [String])
-       -> Teletype v
+       -> Console v
        -> (([String], [String]) -> Arr '[] v ([String], [String]))
        -> Eff '[] ([String], [String])
     go (is, os) (PutStrLn msg) q = q (is, msg : os) ()
@@ -61,26 +61,26 @@ runTeletypePure inputs req =
 -------------------------------------------------------------------------------
                      -- Effectful Interpreter for Deeper Stack --
 -------------------------------------------------------------------------------
-runTeletypeM :: Member IO r => Eff (Teletype ': r) w -> Eff r w
-runTeletypeM (Val x) = return x
-runTeletypeM (E u q) = case decomp u of
-    Right (PutStrLn msg) -> send (putStrLn msg) >> runTeletypeM (qApp q ())
-    Right GetLine        -> send getLine >>= \s -> runTeletypeM (qApp q s)
+runConsoleM :: Member IO r => Eff (Console ': r) w -> Eff r w
+runConsoleM (Val x) = return x
+runConsoleM (E u q) = case decomp u of
+    Right (PutStrLn msg) -> send (putStrLn msg) >> runConsoleM (qApp q ())
+    Right GetLine        -> send getLine >>= \s -> runConsoleM (qApp q s)
     Right ExitSuccess    -> send exitSuccess
-    Left u'              -> E u' (tsingleton (\s -> runTeletypeM (qApp q s)))
+    Left u'              -> E u' (tsingleton (\s -> runConsoleM (qApp q s)))
 
 -------------------------------------------------------------------------------
                      -- Pure Interpreter for Deeper Stack --
 -------------------------------------------------------------------------------
-runTeletypePureM
+runConsolePureM
     :: [String]
-    -> Eff (Teletype ': r) w
+    -> Eff (Console ': r) w
     -> Eff r (Maybe w,([String],[String]))
     -- ^ (Nothing for ExitSuccess, (unconsumed input, produced output))
-runTeletypePureM inputs = f (inputs,[]) where
+runConsolePureM inputs = f (inputs,[]) where
     f
         :: ([String],[String])
-        -> Eff (Teletype ': r) w
+        -> Eff (Console ': r) w
         -> Eff r (Maybe w,([String],[String]))
     f st (Val x) = return (Just x, st)
     f st@(is,os) (E u q) = case decomp u of